Add a simple C CRC-8-CCITT implementation
This adds an 8-bit CRC calculation with the CCITT
polynomial (x^8+x^2+X+1). It uses a 256 byte lookup table
which is pre-calculated using the C pre-preprocessor.
This could have been implemented with constexpr but that
would require c++17 to enable use of std::array in constexpr
functions. Using the pre-processor is more portable for
now.

Code originally comes from a stack overflow answer:
https://stackoverflow.com/a/27843120/6353

Tests:
 - Check against incoming and outgoing logic analyser
   captures of STGAP1AS isolated gate drivers which use this
   algorithm.