Fix CANopen SDO support
This fixes a serious protocol bug in the CANopen SDO
support which meant it was not possible to use other
CANopen tools to read and write values. The command
constants for reading and writing were inverted due to the
standard being confusing. In addition some command
transfer flags were incorrect.

It also fixes a bug introduced in commit d57845 which
prevented the extended dictionary for accessing parameters
by unique ID from operating as intended.

Tests:
 - With a small python script using the "canopen" library verify
    normal operation for reading and writing parameters and
    values:
        print("boost/curkp:", node.sdo.upload(0x2001, 0))
        print("Set boost/curkp to a new value")
        node.sdo.download(0x2000, 0, b'\x00\x04\x00\x00')
        print("boost/curkp:", node.sdo.upload(0x2000, 0))
        print("curkp:", node.sdo.upload(0x2100, 107))
        print("Set curkp to a new value")
        node.sdo.download(0x2100, 107, b'\x01\x02\x00\x00')
        print("curkp:", node.sdo.upload(0x2100, 107))
        print("pot:", node.sdo.upload(0x2107, 0xdf))
 - Verify that an invalid dictionary ID read:
        print("boost/curkp:", node.sdo.upload(0x2001, 0))
 fails with:
   canopen.sdo.exceptions.SdoAbortedError: Code
   0x06020000, Object does not exist
 - Capture CAN packets with Wireshark and verify each
   request and response from the test sequence is decoded
   correctly and matches the expected values and behaviour