Prevent duplicate parameter ids
This introduces an internal enum to params.cpp which contains an
entry for each parameter numbered by its unique id. It is valid
in C/C++ to have multiple entries in an enumeration with the
same number. We want to prevent this so include the enum in a
switch to force the compiler to reject duplicate entries. The
CheckParamIdDuplicates() function is unused so will be omitted
from the final binary.

Tests:
 - Build as part of stm32-sine and verify that manualiq, manualid,
   manualstart, fslipspnt, ampnom, id and iq parameters report
   a clash. Hacking a fix for these clashes causes the compile
   problem to go away.