Enable signed gains when mapping via SDO
It is possible with the serial terminal to
configure negative gain parameters for a can
rx or tx map entry. When reading back via
CAN SDO the map entry's gain value is
incorrect. When adding a can map entry via
SDO the gain is interpreted as a positive value
only.

Tests:
 - Create can map entries via CAN SDO using
   openinverter_can_tool with a mix of :
    - gain values -8388.608, -1.0, 1.0 and 8388.607
    - offset values -128, 0 and 127
 - verify "oic can list" shows the map created
 - verify via serial terminal that "can p" shows
    the same can map
 - check can map operation transmits the
    expected behaviour for a negative gain.