Improve validation when adding can map entries
Validate that the offset and length for a can
map entry will no overflow the resulting CAN
frame. So offset 63/length 1 is valid as is
offset 32/length 32 but offset 60/length 8 is
not. This change affects both serial and SDO.

The length field must be at least 1 bit to make
sense. This was already validated by the SDO
configuration but ignored by serial.

Tests:
 - Verify config of map entries with serial
    terminal fails as documented
 - Verify with a doctored
   openinverter_can_tool without client side
   validation of offset and length that failure is
   reported