Fix CAN rx mapping for negative numbers
Negative numbers are not received properly for can rx maps smaller than 32-bits. To fix this we need to sign extend values so they fit in an int32_t before converting to a float. The sign extension makes no sense for single bits as this would result in -1 being returned rather than 1 when the bit is asserted.

Tests:
 - Create BE and LE 8-bit, 16-bit, 24-bit, 31-bit and
    32-bit mappings
 - Verify with positive and negative numbers that
    SavvyCAN reports the expected value