CAN mapping fixes (#18)
* Enable signed gains when mapping via SDO

It is possible with the serial terminal to
configure negative gain parameters for a can
rx or tx map entry. When reading back via
CAN SDO the map entry's gain value is
incorrect. When adding a can map entry via
SDO the gain is interpreted as a positive value
only.

Tests:
 - Create can map entries via CAN SDO using
   openinverter_can_tool with a mix of :
    - gain values -8388.608, -1.0, 1.0 and 8388.607
    - offset values -128, 0 and 127
 - verify "oic can list" shows the map created
 - verify via serial terminal that "can p" shows
    the same can map
 - check can map operation transmits the
    expected behaviour for a negative gain.

* Improve validation when adding can map entries

Validate that the offset and length for a can
map entry will no overflow the resulting CAN
frame. So offset 63/length 1 is valid as is
offset 32/length 32 but offset 60/length 8 is
not. This change affects both serial and SDO.

The length field must be at least 1 bit to make
sense. This was already validated by the SDO
configuration but ignored by serial.

Tests:
 - Verify config of map entries with serial
    terminal fails as documented
 - Verify with a doctored
   openinverter_can_tool without client side
   validation of offset and length that failure is
   reported

* Send mapped params which overlap first/second words

It is possible to configure a CAN map entry
which spans the internal 32-bit word
boundary used to store can frames. When
sending params to CAN only the bits in the
first 32-bit word are actually sent.

The fix involves detecting the overlap
condition and shifting to the two words. It
uses uint32_t consistently to avoid errors due
to sign-extension when shifting.

The receive path already handled the split
word case.

Tests:
 - Create 16, 24 and 32-bit parameter
    mappings at positions that overlap the 32-
    bit boundary.
 - Use positive and negative values in the
    mapped param
 - Verify with a custom DBC in SavvyCAN that
    the values are interpreted and scales
    appropriately.
- Additionally, add mappings either side of the
   split mapping to verify no data overwrite

* Fix parameters mapped to first 32-bits

Correctly map params of 32-bit in size that fit
entirely within the first word of the CAN
frame.

This fixes an integer overflow in the mask
generation which doesn't matter for 32-bit
ARM but is undefined behaviour and causes
problems when compiled for x86-64.

Tests:
 - Map a negative and positive valued params
    to offset 0, length 32 and verify the CAN
    frame contains the expected value in the
    first 4 bytes.

* Fix validation of big-endian CAN map offsets

In big-endian can mappings the offset
specifies the end bit for the mapping and the
negative length counts bits back. The existing
offset validation only worked for little endian
mappings where the values are all positive.
This changes separates the validation.

Tests:
 - Verify 32-bit mappings work with offsets of
   31, 55, 63
 - Verify 32-bit mappings fail with offsets of 7,
   30 and 64

* Implement big-endian tx can mapping

Big-endian transmit can mappings share the same masking as
little-endian mappings. The bytes are then swapped before
shifting into the appropirate position. The maths is different as
the offset position in big-endian mappings specifies the end bit
offset wheras little-endian it's the start bit.

Tests:
 - Create 8-bit mappings with offset 7 and 39
 - Create 16-bit mappings with offset 23, 39 and 47
 - Create 24-bit mappings with offset 23, 55 and 63
 - Create 32-bit mappings with offset 31, 39, 47, 55, 63
 - Test each mapping with positive and negative values
 - Verify the output in SavvyCAN with equivalent Message/Signal

* Fix CAN rx mapping for negative numbers

Negative numbers are not received properly for can rx maps smaller than 32-bits. To fix this we need to sign extend values so they fit in an int32_t before converting to a float. The sign extension makes no sense for single bits as this would result in -1 being returned rather than 1 when the bit is asserted.

Tests:
 - Create BE and LE 8-bit, 16-bit, 24-bit, 31-bit and
    32-bit mappings
 - Verify with positive and negative numbers that
    SavvyCAN reports the expected value