Create unsigned CAN receive tests
This adds the missing unit tests to cover the receive
mapping of CAN values where the value is larger than
the limit of a signed value. This demonstrates that the
unsigned mapping is working as expected.

The CanMap code has a conditional define to allow
building of libopeninv with signed receive support.
Extend the unit test Makefile to support building with
this and retaining the existing signed receive mapping
tests.

Extend the GitHub Action to build both variants.

Tests:
 - Build with CAN_SIGNED=0 and verify that 32-bit
    receive tests are broken
 - Build with CAN_SIGNED=1 and verify existing tests
    pass