Fix unsigned CAN receive for 32-bit values
With a receive CAN mapping for a 32-bit value if the
value being received is > 2^31 then it is converted into
a negative number rather than the large positive
number that would be expected. The fix is to avoid
storing the incoming value in a signed type when
building the normal unsigned version of libopeninv.

Tests:
 - Build and run unit tests with CAN_SIGNED=1 and
   CAN_SIGNED=0 with no errors