Fix unsigned CAN receive mapping for 32-bit values and associated unit testing (#22)
* Import unit tests from stm32-sine

Pull in stm32-sine unit tests from commit 1b43e4c.

Remove stm32-sine specific tests.

Add minimal param_prj.h and hwdefs.h and missing
libopencm3 stubs to get the unit tests to link and run.

Construct a GitHub action to build just the unit tests
on Linux

* Create unsigned CAN receive tests

This adds the missing unit tests to cover the receive
mapping of CAN values where the value is larger than
the limit of a signed value. This demonstrates that the
unsigned mapping is working as expected.

The CanMap code has a conditional define to allow
building of libopeninv with signed receive support.
Extend the unit test Makefile to support building with
this and retaining the existing signed receive mapping
tests.

Extend the GitHub Action to build both variants.

Tests:
 - Build with CAN_SIGNED=0 and verify that 32-bit
    receive tests are broken
 - Build with CAN_SIGNED=1 and verify existing tests
    pass

* Fix unsigned CAN receive for 32-bit values

With a receive CAN mapping for a 32-bit value if the
value being received is > 2^31 then it is converted into
a negative number rather than the large positive
number that would be expected. The fix is to avoid
storing the incoming value in a signed type when
building the normal unsigned version of libopeninv.

Tests:
 - Build and run unit tests with CAN_SIGNED=1 and
   CAN_SIGNED=0 with no errors