Fix LIN checksum calculation (#33)
There is a typo in the LIN checksum calculation that
introduces a data dependency bug. According to
Section 2.3.1.5 Checksum of the LIN Protocol
Specification Revision 2.1 the checksum should
check if the value is greater than or equal to 256
before subtracting 255.

Add a set of unit tests with representative LIN data.
Only TestSyntheticLinStatusResponse1() actually
exhibits the bug.

Tests:
 - Run new unit tests
 - Verify VWCoolantHeater simulator built with ESP32
   and https://github.com/gicking/LIN_slave_portable_Arduino
   communicates data reliably with ZombieVerter VCU