Allow listing of errors via CANopen SDO
This exports the list of errors and time at which they
occurred as two CANopen SDO arrays. The sub-
index is used as an index into the list of errors. Any
invalid indices return zero.

Clients can use the "lasterr" parameter database
spot-value which is present in all libopeninv projects
to map the error number to a string.

Tests:
 - Verify with updated OpenInverter CAN Tool that
   errors can be listed.
 - Initiate an new error after a period of time and
   ensure the list tallies with the elapsed wall-clock
   time.
 - Check CANopen SDO decodes as expected in
   Wireshark